+++ libraries/PT2314/CHANGELOG.md
@@ -6,10 +6,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.2.0] - 2023-12-08
+- refactor API, begin()
+- update readme.md
+- added **void setAttn(uint8_t attn)**
+- removed setAttn(left, right).
+- minor edits
+
+----
+
 ## [0.1.2] - 2023-09-24
 - add Wire1 support for ESP32
 - update readme.md
-
 
 ## [0.1.1] - 2023-08-01
 - add **PT7313** derived class